P/Invoke fixes
This makes the P/Invoke interface work again.

No idea is this will be useful, but it really does put things back to working in a pure DNA built on to run on the command line.

I thin that adding in the extra params was the main mistake. This makes the P/Invoke fall in to a little heap on the floor. I think making an extra attribute or calling convention for the C# sire would be a better method for gettig this working.